1955 | Separate overflow test to prevent open fail on macOS
